Both model done (v1.0)
