+++ models/generate_dl_model.py
@@ -1,139 +1,150 @@
 """
 File to generate the DL model which I'll use to predict the genre of a synopsis
 """
+
 import torch
 import torch.nn as nn
-import torch.optim as optim
-from sklearn.metrics import accuracy_score
-import numpy as np
 from torch.utils.data import Dataset, DataLoader
+import polars as pl
+from pathlib import Path
+from typing import List, Tuple, Dict
+import logging
+from generate_ml_model import load_data
+from sklearn.preprocessing import LabelEncoder
 
-PAD_IDX = 0
-UNK_IDX = 1
-MAX_LEN = 100
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+PROJECT_ROOT = Path(__file__).parent.parent
+MODEL_SAVE_PATH = PROJECT_ROOT / "models" / "dl_model.pkl"
+EMBEDDING_DIM = 50
 
 
-def load_glove_embeddings(glove_path: str, word2idx : dict, embedding_dim=100):
-    embeddings = np.random.uniform(-0.25, 0.25, (len(word2idx), embedding_dim))
-    embeddings[word2idx['PAD']] = np.zeros(embedding_dim)
+def create_vocabulary(tokens: List[List[str]]) -> Tuple[Dict[str, int], Dict[int, str]]:
+    word2idx = {}
+    idx2word = {}
+    for sentence in tokens:
+        for word in sentence:
+            if word not in word2idx:
+                idx2word[len(word2idx)] = word
+                word2idx[word] = len(word2idx)
+    return word2idx, idx2word
 
-    with open(glove_path, 'r', encoding="utf-8") as f:
-        for line in f:
-            parts = line.strip().split()
-            word = parts[0]
-            if word in word2idx:
-                vect = np.array(parts[1:], dtype=np.float32)
-                embeddings[word2idx[word]] = vect
-    
-    return torch.tensor(embeddings, dtype=torch.float)
 
-def text_to_sequence(text : str, word2idx: dict, max_len=MAX_LEN):
-    tokens = text.split() # We only use split since we preprocess data ourselves
-    seq = [word2idx.get(tok, UNK_IDX) for tok in tokens]
-    if len(seq) < MAX_LEN:
-        seq += [PAD_IDX] * (max_len - len(seq))
-    else:
-        seq = seq[:max_len]
-    return seq
-
-class TextDataset(Dataset):
-    def __init__(self, texts, labels, word2idx, max_len=MAX_LEN):
-        self.texts = texts
+class GenreDataset(Dataset):
+    def __init__(self, sequences, labels):
+        self.sequences = sequences
         self.labels = labels
-        self.word2idx = word2idx
-        self.max_len = max_len
 
     def __len__(self):
-        return len(self.texts)
-    
-    def __getitem__(self, idx):
-        seq = text_to_sequence(self.texts[idx], self.word2idx, self.max_len)
-        label = self.labels[idx]
-        return torch.tensor(seq, dtype=torch.long), torch.tensor(label, dtype=torch.long)
+        return len(self.sequences)
 
-class TextClassifier(nn.Module):
-    def __init__(self, embedding_matrix, hidden_dim, output_dim, freeze_embeddings=True):
+    def __getitem__(self, idx):
+        return self.sequences[idx], self.labels[idx]
+
+
+class GenreClassifier(nn.Module):
+    def __init__(
+        self,
+        vocab_size,
+        hidden_dim: int,
+        pad_idx: torch.Tensor,
+        output_dim=1,
+        embedding_dim=EMBEDDING_DIM,
+    ):
         super().__init__()
-        vocab_size, embeddings_dim = embedding_matrix.shape
-        self.embedding = nn.Embedding.from_pretrained(embedding_matrix, freeze=freeze_embeddings, padding_idx=PAD_IDX)
-        self.lstm = nn.LSTM(embeddings_dim, hidden_dim, batch_first=True, bidirectional=True)
+        self.embedding = nn.Embedding(vocab_size, embedding_dim, padding_idx=pad_idx)
+        self.lstm = nn.LSTM(
+            embedding_dim, hidden_dim, batch_first=True, bidirectional=True
+        )
+        self.dropout = nn.Dropout(0.3)
         self.fc = nn.Linear(hidden_dim * 2, output_dim)
-        self.dropout = nn.Dropout(0.5)
-    
+
     def forward(self, x):
         embedded = self.embedding(x)
-        lstm_out = self.lstm(embedded)
+        lstm_out, _ = self.lstm(embedded)
         pooled = torch.mean(lstm_out, dim=1)
-        dropped = self.dropout(pooled)
-        out = self.fc(dropped)
-        return out
-    
-def train_epochs(model : TextClassifier, dataloader, criterion, optimizer : optim.Optimizer, device : torch.device):
-    model.train()
-    epoch_loss = 0
-    preds = []
-    trues = []
+        out = self.dropout(pooled)
+        return self.fc(out)
 
-    for inputs, labels in dataloader:
-        inputs, labels = inputs.to(device), labels.to(device)
+
+def train_epoch(
+    model: GenreClassifier,
+    dataloader: DataLoader,
+    criterion,
+    optimizer,
+    device: torch.device,
+):
+    model.train()
+    total_loss, total_correct = 0, 0
+    for x_batch, y_batch in dataloader:
+        x_batch, y_batch = x_batch.to(device), y_batch.to(device)
         optimizer.zero_grad()
-        outputs = model(inputs)
-        loss = criterion(outputs, labels)
+        preds = model(x_batch)
+        loss = criterion(preds, y_batch)
         loss.backward()
         optimizer.step()
-
-        epoch_loss += loss.item() * inputs.size(0) 
-        preds.extend(torch.argmax(outputs, dim=1).cpu().numpy())
-        trues.extend(labels.cpu().numpy())
-
-    acc = accuracy_score(trues, preds)
-    return epoch_loss / len(dataloader.dataset), acc
-
-def eval_epoch(model : TextClassifier, dataloader, criterion, device : torch.device):
-    model.eval()
-    epoch_loss = 0
-    preds = []
-    trues = []
-
-    with torch.no_grad():
-        for inputs, labels in dataloader:
-            inputs, labels = inputs.to(device), labels.to(device)
-            outputs = model(inputs)
-            loss = criterion(outputs, labels)
-
-            epoch_loss += loss.item() * inputs.size(0)
-            preds.extend(torch.argmax(outputs, dim=1).cpu().numpy())
-            trues.extend(labels.cpu().numpy())
-
-    acc = accuracy_score(trues, preds)
-    return epoch_loss / len(dataloader.dataset), acc
+        total_loss += loss.item() * x_batch.size(0)
+        total_correct += (preds.argmax(1) == y_batch).sum().item()
+    return total_loss / len(dataloader.dataset), total_correct / len(dataloader.dataset)
 
 
-if __name__ == ("__main__"):
-    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+if __name__ == "__main__":
 
+    """
+    Model preparation
+    """
+    logger.info("Model preparation start..")
+    X_train, X_test, y_train, y_test = load_data()
 
-    # Load your embeddings
-    embedding_matrix = load_glove_embeddings("path/to/glove.6B.100d.txt", word2idx, 100)
+    # We can tokenize with split since X_train is already preprocessed data
+    tokens = [sentence.split() for sentence in X_train]
 
-    # Prepare datasets and dataloaders
-    train_dataset = TextDataset(train_texts, train_labels, word2idx)
-    val_dataset = TextDataset(val_texts, val_labels, word2idx)
+    word2idx, idx2word = create_vocabulary(tokens)
+
+    indices = [[word2idx[word] for word in sentence] for sentence in tokens]
+
+    padded_indices = nn.utils.rnn.pad_sequence(
+        [torch.LongTensor(sentence) for sentence in indices], batch_first=True
+    )
+
+    label_encoder = LabelEncoder()
+
+    y_train_encoded = label_encoder.fit_transform(y_train)
+
+    train_labels = torch.tensor(y_train_encoded, dtype=torch.long)
+
+    train_dataset = GenreDataset(padded_indices, train_labels)
+
     train_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
-    val_loader = DataLoader(val_dataset, batch_size=64)
-
-    # Build model
-    model = TextClassifier(embedding_matrix, hidden_dim=128, output_dim=num_classes)
-    model.to(device)
-
+    logger.info("Model preparation done.")
+    """
+    Model training
+    """
+    device = torch.device("cude" if torch.cuda.is_available() else "cpu")
+    num_classes = len(label_encoder.classes_)
+    model = GenreClassifier(
+        vocab_size=len(word2idx),
+        embedding_dim=EMBEDDING_DIM,
+        hidden_dim=128,
+        output_dim=num_classes,
+        pad_idx=0,
+    ).to(device)
+    logger.info("Model instantiation done.")
     criterion = nn.CrossEntropyLoss()
-    optimizer = optim.Adam(model.parameters(), lr=1e-3)
-
-    num_epochs = 10
-    for epoch in range(num_epochs):
-        train_loss, train_acc = train_epoch(model, train_loader, criterion, optimizer, device)
-        val_loss, val_acc = eval_epoch(model, val_loader, criterion, device)
-        print(f"Epoch {epoch+1}: Train Loss {train_loss:.3f} Acc {train_acc:.3f} | Val Loss {val_loss:.3f} Acc {val_acc:.3f}")
-
-    # Save the model
-    torch.save(model.state_dict(), "models/dl_model.pth")
+    optimizer = torch.optim.Adam(model.parameters(), lr=1e-3)
+    logger.info("Training phase start..")
+    for epoch in range(50):
+        train_loss, train_acc = train_epoch(
+            model=model,
+            dataloader=train_loader,
+            criterion=criterion,
+            optimizer=optimizer,
+            device=device,
+        )
+        logger.info(
+            f"Epoch {epoch + 1}: Loss={train_loss:.4f}, Accuracy={train_acc:.4f}"
+        )
+    logger.info("Training phase end..")
+    torch.save(model.state_dict(), MODEL_SAVE_PATH)
+    logger.info("Deep Learning model savec.")
